+++ polymerist/mdtools/openfftools/partialcharge/rescharge/calculation.py
@@ -14,9 +14,9 @@ from openmm.unit import elementary_charge
 
 from .rctypes import ChargesByResidue
 from .redistribution import ChargeRedistributionStrategy, UniformDistributionStrategy
-from ....monomers.repr import MonomerGroup
-from ....maths.statistics import Accumulator
-from ....genutils.decorators.functional import optional_in_place
+from .....genutils.decorators.functional import optional_in_place
+from .....maths.statistics import Accumulator
+from .....polymers.monomers.repr import MonomerGroup
 
 
 # FUNCTIONS FOR RESIDUE CHARGE CALCULATION
