+++ polymerist/polymers/estimation.py
@@ -5,7 +5,7 @@ from rdkit import Chem
 
 from .exceptions import InsufficientChainLengthError
 from ..genutils.iteration import iter_len
-from ..monomers.repr import MonomerGroup
+from ..polymers.monomers.repr import MonomerGroup
 from ..rdutils.bonding.portlib import get_num_ports
