+++ polymerist/polymers/monomers/conversion.py
