+++ polymerist/polymers/monomers/specification.py
